abstract pool status items
